[Refactor] #1370 Moved throw-util.cpp/h to throw-execution.cpp/h
