+++ src/cmd-item/cmd-throw.cpp
@@ -4,7 +4,7 @@
  * @author Hourier
  */
 #include "cmd-item/cmd-throw.h"
-#include "action/throw-util.h"
+#include "object-use/throw-execution.h"
 #include "core/player-redraw-types.h"
 #include "game-option/special-options.h"
 #include "inventory/inventory-slot-types.h"
